// bun-sistema-colas/src/types/queue-message-state.ts
import type { Ticket } from './ticket';

export interface QueueMessageState {
  lastTicketNumbers: {
    A: number;
    P: number;
  };
  pendingTotal: {
    normal: number;
    preferential: number;
    combined: number;
  };
  activeByDesk: Record<number, Ticket | undefined>;
  recentlyServed: Ticket[]; // último 8
}